requirements.txt
transformers
torch
flask
requests
pillow
